More structure rework
Added an expanded class for Flask to add another decorator (endpoint) which will
register a route as an endpoint (meaning it is a specific entry point into the app,
not relying on dynamic routes given, will have examples later).  Also changed the
__main__.py file to live outside the module, was breaking the debug mode of flask.
Expanded the datatype decorator to take an optional argument that is the template,
the code to make that argument optional is ugly, need to fix it up some.  Added
a base template and an example index one (uses the base), need to write these out
as time goes along.