Incremental commit
This includes a basic expansion of the logging system (the wrapper handles creating
the extra information and setup of the log messages), the inclusion of html.py,
which is just used to serve HTML pages for forms that can be done via AJAX.  The
database errors were expanded, now print as nice string representations.  Wrote in
some rudimentary roles for users (need to work this in better).  Expanded the User
set for the database side (not fully tested).  Just a lot of random fixes.  I really
need to start committing more incrementally with different tasks, just too much
fixing going on right now.  Tomorrow...