Few small fixes made
Logger messages don't need to be run through format (the logger does it itself).
The roles in the permissions decorator are both lists, so an in will always fail.
Wrote the intersect helper function which will return a boolean if the two lists
have any intersects (this could be fixed by using sets).