Draft of the JS library
This is just a draft (untested/used) of a frontend javascript library that will be
an example of how to consume the API.  I will try to comment it well and explain
the various things.  Right now, only works if jQuery is already loaded on the page,
need to write the stuff to work without jQuery.