Tests now have a better docstring
The first line of the docstring for a test is used to describe it, changed from the
ugly SuiteName::method_name to a short description of the test (the longer
description is the rest of the docstring).  Also fixed some dumb code in the
PlayerTest stuff.  The created player was always the same (the passed in player
object was entirely ignored).