Reorganization leading up to adding quests
Moved the filter_keys function to be a first order for the module.  Made the modify
player route use PUT instead of POST (modify should be PUT).  Moved the
create_player test method to the TestBase.  This should have been broken up, but I
was too involved in prepping the quests stuff.  :(