Implemented first pass of quests
Quests currently support admins or root creating quests, retrieving the list,
marking that a player has completed a quest (with basic application of the quest's
rewards to the player).  Tests have been written, little coverage for now.  Now, if
the retrieval of the single player document (get_player) has a quest key in the
query string, it will include a list of the quests the player has completed.