Added static serving in DEBUG mode
Will serve static files in DEBUG mode using the settings (if set) in the settings
file.  The folder key points to the path to look for the static files (../lib is
the template default which points at the repo_root/lib) and the path key tells what
prefix on the URL will flag it as a static file (/s is the default, so /s/rpg.js
would serve /lib/rpg.js).  Static files are not served in non DEBUG mode, a proper
web server should handle sending static files.