JS library works so far
The JS library was tested (using the new static serving).  It works for the
currently written parts of the library.  The jQuery interaction could use some more
love (I don't think the error handling is working correctly).  This needs a lot of
love for designing the usage.  Maybe that will be on the plate soon.